+++ lib/src/alert_dialog/show_confirmation_dialog.dart
@@ -14,6 +14,9 @@ import 'package:meta/meta.dart';
 /// for performance optimization.
 /// if [initialSelectedActionKey] is set, corresponding action is selected
 /// initially. This works only for Android style.
+/// [toggleable] option is only supported in material dialog.
+/// When set to `false`, the selected action cannot be unselected by tapping.
+/// The default value is `true`.
 @useResult
 Future<T?> showConfirmationDialog<T>({
   required BuildContext context,
@@ -29,9 +32,11 @@ Future<T?> showConfirmationDialog<T>({
   bool useRootNavigator = true,
   bool shrinkWrap = true,
   bool fullyCapitalizedForMaterial = true,
-  WillPopCallback? onWillPop,
+  bool canPop = true,
+  PopInvokedCallback? onPopInvoked,
   AdaptiveDialogBuilder? builder,
   RouteSettings? routeSettings,
+  bool toggleable = true,
 }) {
   void pop({required BuildContext context, required T? key}) => Navigator.of(
         context,
@@ -60,7 +65,9 @@ Future<T?> showConfirmationDialog<T>({
               contentMaxHeight: contentMaxHeight,
               shrinkWrap: shrinkWrap,
               fullyCapitalized: fullyCapitalizedForMaterial,
-              onWillPop: onWillPop,
+              canPop: canPop,
+              onPopInvoked: onPopInvoked,
+              toggleable: toggleable,
             );
             return builder == null ? dialog : builder(context, dialog);
           },
@@ -73,7 +80,8 @@ Future<T?> showConfirmationDialog<T>({
           actions: actions.convertToSheetActions(),
           style: style,
           useRootNavigator: useRootNavigator,
-          onWillPop: onWillPop,
+          canPop: canPop,
+          onPopInvoked: onPopInvoked,
           builder: builder,
           routeSettings: routeSettings,
         );
@@ -92,7 +100,9 @@ class _ConfirmationMaterialDialog<T> extends StatefulWidget {
     @required this.contentMaxHeight,
     required this.shrinkWrap,
     required this.fullyCapitalized,
-    required this.onWillPop,
+    required this.canPop,
+    required this.onPopInvoked,
+    required this.toggleable,
   });
 
   final String title;
@@ -105,7 +115,9 @@ class _ConfirmationMaterialDialog<T> extends StatefulWidget {
   final double? contentMaxHeight;
   final bool shrinkWrap;
   final bool fullyCapitalized;
-  final WillPopCallback? onWillPop;
+  final bool canPop;
+  final PopInvokedCallback? onPopInvoked;
+  final bool toggleable;
 
   @override
   _ConfirmationMaterialDialogState<T> createState() =>
@@ -130,8 +142,8 @@ class _ConfirmationMaterialDialogState<T>
     final cancelLabel = widget.cancelLabel;
     final okLabel = widget.okLabel;
     final message = widget.message;
-    return WillPopScope(
-      onWillPop: widget.onWillPop,
+    return PopScope(
+      canPop: widget.canPop,
       child: Dialog(
         child: Column(
           mainAxisSize: MainAxisSize.min,
@@ -172,7 +184,10 @@ class _ConfirmationMaterialDialogState<T>
                   children: widget.actions
                       .map(
                         (action) => RadioListTile<T>(
-                          title: Text(action.label),
+                          title: Text(
+                            action.label,
+                            style: action.textStyle,
+                          ),
                           value: action.key,
                           groupValue: _selectedKey,
                           onChanged: (value) {
@@ -180,7 +195,7 @@ class _ConfirmationMaterialDialogState<T>
                               _selectedKey = value;
                             });
                           },
-                          toggleable: true,
+                          toggleable: widget.toggleable,
                         ),
                       )
                       .toList(),