+++ lib/src/modal_action_sheet/material_modal_action_sheet.dart
@@ -1,5 +1,6 @@
 import 'package:adaptive_dialog/src/action_callback.dart';
 import 'package:flutter/material.dart';
+
 import 'modal_action_sheet.dart';
 
 class MaterialModalActionSheet<T> extends StatelessWidget {
@@ -10,7 +11,8 @@ class MaterialModalActionSheet<T> extends StatelessWidget {
     this.title,
     this.message,
     this.materialConfiguration,
-    this.onWillPop,
+    required this.canPop,
+    required this.onPopInvoked,
   });
 
   final ActionCallback<T> onPressed;
@@ -18,7 +20,8 @@ class MaterialModalActionSheet<T> extends StatelessWidget {
   final String? title;
   final String? message;
   final MaterialModalActionSheetConfiguration? materialConfiguration;
-  final WillPopCallback? onWillPop;
+  final bool canPop;
+  final PopInvokedCallback? onPopInvoked;
 
   @override
   Widget build(BuildContext context) {
@@ -62,7 +65,7 @@ class MaterialModalActionSheet<T> extends StatelessWidget {
             a.label,
             style: TextStyle(
               color: color,
-            ),
+            ).merge(a.textStyle),
           ),
           onTap: () => onPressed(a.key),
         );
@@ -85,8 +88,9 @@ class MaterialModalActionSheet<T> extends StatelessWidget {
               children: children,
             ),
           );
-    return WillPopScope(
-      onWillPop: onWillPop,
+    return PopScope(
+      canPop: canPop,
+      onPopInvoked: onPopInvoked,
       child: body,
     );
   }